rename mod `erc20_token_activation` to `eth_tokens_activation`
