+++ mm2src/coins_activation/src/eth_tokens_activation.rs
